Update patch set 1
Patch Set 1: Code-Review-1 Verified-1

The problem is not the crypto support. The problem is that recovery is booting with SELinux enforcing and the sepolicy in Android 5.0 is too restrictive. I've come up with 3 different methods to set permissive in recovery that I've pushed up to Gerrit. I'm favoring the approach in https://gerrit.omnirom.org/#/c/10924/ and https://gerrit.omnirom.org/#/c/10925/

The decrypt support for 5.0 L decrypt won't work on the Nexus 4 unless / until we update the vendor tree to copy the right vendor files into the TWRP ramdisk.

Patch-set: 1