Only allow toolbox exec where /system exec was already allowed.
When the toolbox domain was introduced, we allowed all domains to exec it
to avoid breakage.  However, only domains that were previously allowed the
ability to exec /system files would have been able to do this prior to the
introduction of the toolbox domain.  Remove the rule from domain.te and add
rules to all domains that are already allowed execute_no_trans to system_file.
Requires coordination with device-specific policy changes with the same Change-Id.

Based on:
https://android.googlesource.com/device/asus/flo/+/078640e52120e9f86e83104de1e90f7b6608bc82

Change-Id: Ifcd21fb63ca9eb0052bab5561e6897d6b984a8de